+++ src/Auth/UseConfigInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace YdbPlatform\Ydb\Auth;
+
+interface UseConfigInterface
+{
+    public function setYdbConnectionConfig(array $config);
+}