+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 * added StaticAuthentication
+* added query timeout and canceled params
 
 ## 1.11.0
 * added query stats